+++ src/main/java/com/bruno/cursomc/services/ClienteService.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort.Direction;
@@ -65,15 +64,27 @@ public class ClienteService {
 		return repo.save(newObj); // salva o newObj
 	}
 
+	// Esse método foi comentada porque excluia categoria mesmo tendo produtos relacionados.
+//	public void delete(Integer id) {
+//		// caso id não existe, irá executar uma exceção do método find
+//		find(id);
+//		try {
+//			repo.deleteById(id);
+//		}
+//		// capturar a excessão de Integridade, lançando a excessão personalizada
+//		catch (DataIntegrityViolationException e) {
+//			throw new DataIntegrityException("Não é possível excluir porque há pedidos relacionadas");
+//		}
+//	}
 	
-	public void delete(Integer id) {
-		// caso id não existe, irá executar uma exceção do método find
-		find(id);
-		try {
+	public void delete(Integer id){
+
+		Cliente cliente = find(id);
+		
+		if (cliente.getPedidos().isEmpty()) {
 			repo.deleteById(id);
 		}
-		// capturar a excessão de Integridade, lançando a excessão personalizada
-		catch (DataIntegrityViolationException e) {
+		else {
 			throw new DataIntegrityException("Não é possível excluir porque há pedidos relacionadas");
 		}
 	}
